Replace RenderFullScreen with top layer
This patch makes the Fullscreen API use the top layer, as per the latest
Fullscreen spec:
http://fullscreen.spec.whatwg.org/

Some more details:
- HTMLPlugInElement now retains its widget during reattach. Otherwise,
a video would need to be reloaded when it enters fullscreen.
- The placeholder renderer that was in RenderFullScreen is removed. The
spec doesn't mention such a mechanism and it's better to avoid the
complexity if possible. It was originally added to prevent a flash of a
collapsed page when exiting fullscreen:
https://bugs.webkit.org/show_bug.cgi?id=61897
It's possible we still need it. Either way, it has to be moved from
RenderFullScreen.

Intent to ship:
https://groups.google.com/a/chromium.org/forum/#!topic/blink-dev/yJUanvg7d40

BUG=240576,246077

Review URL: https://codereview.chromium.org/139743005

git-svn-id: svn://svn.chromium.org/blink/trunk@165710 bbb929c8-8fbe-4397-9dbb-9b2b20218538