This CL cleans up import-w3c-tests so that the default behavior matches what we want.
Namely, in the normal case, you want to import the tests in place in a checkout (i.e., you would cd to LayoutTests/w3c/csswg-test and just run the script), rather than copying things from one place to another.

This patch also adds the 'W3CImportExpectations' so that we can track which directories to blacklist (making it clearer which ones are being imported, plus getting new directories as they are added for free).

We leave the old behavior in place in case someone still wants to do a manual import.

R=ojan@chromium.org
BUG=242329

Review URL: https://codereview.chromium.org/130903007

git-svn-id: svn://svn.chromium.org/blink/trunk@165968 bbb929c8-8fbe-4397-9dbb-9b2b20218538