Blocks that create a new block formatting context are not self-collapsing.
Per http://www.w3.org/TR/CSS2/box.html#collapsing-margins the top and bottom margins of a block that has no content and otherwise no height of its own can only collapse together if the box doesn't establish a block formatting context.

We already have a function that detects when a block is creating a new block formatting context, so rename it and use it to detect when we shouldn't collapse margins.

The newly renamed 'createsNewBlockFormattingContext' function needs to learn about other situations that create block formatting contexts (such as column-spans and regions) but that will be done in separate patches.

BUG=321295

Review URL: https://codereview.chromium.org/144213004

git-svn-id: svn://svn.chromium.org/blink/trunk@166009 bbb929c8-8fbe-4397-9dbb-9b2b20218538