Enable WebSocket new impl behind a runtime enabled feature.
Use NewWebSocketChannelImpl when Blink runtime enabled feature
'ExperimentalWebSocket' is enabled.

We enabled NewWebSocketChannelImpl behind a chromium-side flag at [A], but
it was reverted because it leads to crashes on chrome with the flag
turned on (the chromium-side flag was used previously for another feature
which is stable now and some users turns it on). Hence we change the strategy.

We are going to:

1: Land this CL.
2: Turn the blink runtime enabled feature on on LayoutTests and
   run LayoutTests for both the old impl and the new impl.
3: Stabilize the impl.
4: Land [B] to bind the chromium-side flag and the runtime enabled feature.
5: Remove Settings::ExperimentalWebSocketEnabled.

The chromium-side flag will take effect when [B] is landed.

[A]: https://src.chromium.org/viewvc/blink?view=rev&revision=165637
[B]: https://codereview.chromium.org/138443010/

BUG=275458
R=ricea@chromium.org

Review URL: https://codereview.chromium.org/149363003

git-svn-id: svn://svn.chromium.org/blink/trunk@166092 bbb929c8-8fbe-4397-9dbb-9b2b20218538