Report touch rect for whole non-composited scrollable layer with a child touch handler.
Currently, touch hit rects for an element inside a non-composited scrollable
layer are generated relative to the nearest composited ancestor. On scrolling,
the layer is repainted but the touch rects aren't recomputed. It would be
prohibitive to recompute the rects on every scroll so instead increase the rects
to cover the bounding box of the outermost non-composited scrollable ancestor
which would guarantee that the element is always inside a hit rect.

BUG=261307

Review URL: https://codereview.chromium.org/149043002

git-svn-id: svn://svn.chromium.org/blink/trunk@166260 bbb929c8-8fbe-4397-9dbb-9b2b20218538