Fix scroll restore when doing a same-document back navigation.
http://src.chromium.org/viewvc/blink?revision=164641&view=revision broke it.

There's an old, nasty hack in WebFrameImpl::currentHistoryItem() where we
save state to the current history item before we return it to the embedder.
r164641 slightly expanded the case where we do so, causing us to clobber
scroll position before restoring it. I'm suspicious that saving the scroll
state before returning it is unnecessary, or at the very least is a
workaround for a case where we aren't saving the state earlier.

Ergo, remove the hack, that we may fix the current issue
and see what monsters lurk beneath.

BUG=339822
TEST=http/tests/navigation/same-document-scroll-position.html (revised)

Review URL: https://codereview.chromium.org/137803010

git-svn-id: svn://svn.chromium.org/blink/trunk@166428 bbb929c8-8fbe-4397-9dbb-9b2b20218538